feature(ai-generator) : working in progress for image generator and linking it with the front and back end
